+++ GuaranteedRate.Sextant/Properties/AssemblyInfo.cs
@@ -36,5 +36,5 @@ using System.Runtime.InteropServices;
 // by using the '*' as shown below:
 // [assembly: AssemblyVersion("1.0.*")]
 
-[assembly: AssemblyVersion("16.2.5.3")]
-[assembly: AssemblyFileVersion("16.2.5.3")]
+[assembly: AssemblyVersion("16.2.6.0")]
+[assembly: AssemblyFileVersion("16.2.6.0")]